查询添加

+++ src/main/java/com/klw/oa/dao/UserMapper.java
@@ -3,13 +3,13 @@ package com.klw.oa.dao;
 import com.klw.oa.entity.User;
 
 public interface UserMapper {
-    int deleteByPrimaryKey(Long userId);
+    int deleteByPrimaryKey(Integer userId);
 
     int insert(User record);
 
     int insertSelective(User record);
 
-    User selectByPrimaryKey(Long userId);
+    User selectByPrimaryKey(Integer userId);
 
     int updateByPrimaryKeySelective(User record);
